requirements.txt
python-mpv-jsonipc>=1.0.0
ssdpy>=0.4.1
pyyaml>=6.0
netifaces>=0.11.0
